refactor: move files to folders by type and update imports
