+++ index.ts
@@ -1,0 +1,1 @@
+export { ContentfulClient } from './src/cms/contentful';
